requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.5
dj-database-url==3.1.2
Django==6.0.3
idna==3.11
psycopg2-binary==2.9.11
python-dotenv==1.2.2
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3
whitenoise==6.12.0
yt-dlp==2026.3.3
gunicorn==23.0.0
